Refactor: 좋아요 수정, 기본이미지 s3 적용
