+++ src/containers/goods/detail/DetailInfo.tsx
@@ -1,12 +1,7 @@
 'use client'
 
-import { useRouter } from 'next/navigation'
-import { useSession } from 'next-auth/react'
-import { useState } from 'react'
-import { LiaHeart, LiaHeartSolid } from 'react-icons/lia'
-import { useToastStore } from '@/components/Toast/store'
+import LikeBtn from '@/components/Btn/LikeBtn'
 import { GoodsDetailType } from '@/types/goodsType'
-import { addLike, deleteLike, getLikeWhether } from '@/utils/mainApiActions'
 import Timer from './Timer'
 
 export default function DetailInfo({
@@ -14,11 +9,6 @@ export default function DetailInfo({
 }: {
   goodsDetail: GoodsDetailType
 }) {
-  const { data: session } = useSession()
-  const router = useRouter()
-  const [isLiked, setIsLiked] = useState<boolean>(false)
-  const { showToast } = useToastStore()
-
   const wishTrade = [
     {
       id: 0,
@@ -34,43 +24,11 @@ export default function DetailInfo({
     },
   ]
 
-  const handleLike = async () => {
-    if (!session) {
-      router.push(`/login?callbackUrl=${window.location.href}`)
-    } else {
-      const whether = await getLikeWhether(goodsDetail.goodsCode)
-
-      if (whether.status === 200) {
-        if (isLiked) {
-          const data = await deleteLike(goodsDetail.goodsCode)
-          if (data.status === 200) {
-            setIsLiked(!isLiked)
-          }
-        } else {
-          const data = await addLike(goodsDetail.goodsCode)
-          if (data.status === 200) {
-            setIsLiked(!isLiked)
-          }
-        }
-      } else if (whether.status === 401) {
-        showToast('로그인이 필요한 서비스입니다.')
-      } else {
-        showToast(whether.message)
-      }
-    }
-  }
-
   return (
     <div className="mt-[35px] px-[20px]">
-      <div className="flex justify-between mt-[15px] items-center">
+      <div className="relative flex justify-between mt-[15px] items-center">
         <p className="text-[20px]">{goodsDetail.name}</p>
-        <button type="button" onClick={() => handleLike()}>
-          {isLiked ? (
-            <LiaHeartSolid className="w-[30px] h-[32px] ml-[13px] text-red-500" />
-          ) : (
-            <LiaHeart className="w-[30px] h-[32px] ml-[13px] text-slate-400" />
-          )}
-        </button>
+        <LikeBtn goodsCode={goodsDetail.goodsCode} />
       </div>
       <div className="flex justify-between items-end">
         <div className="mt-[15px] leading-[30px]">
@@ -80,7 +38,6 @@ export default function DetailInfo({
             <span className="text-[20px] font-normal">원</span>
           </p>
         </div>
-
         <div className="flex flex-col mb-2">
           <p className="pr-[15px] text-[#989898]">거래는</p>
           {goodsDetail.wishTradeType === 0 && (
